serialize(): create temporary file at current dir
to avoid error of `os.replace` when src and dst are on different
filesystems in unix system.

The original (commented code) did create the temporary file
at current directory,
because that time I am using a system
where `/tmp` and `~` are on different filesystems.
And the default behavior (create temporary file at temporary directory)
failed.
Then I changed the code to create temporary file at current directory
but forget to write a comment about this.

Then in later refactoring I made a typo and forgot the dir parameter.
Unfortunately I was on a system
where `/tmp` and `~` are on same filesystem.
So this bug does not caught by using the script.

Thanks @F0nebula to report this bug.

https://www.v2ex.com/member/F0nebula